Replaced hard-coded string IDs with constants
I have no idea if this is the right approach. I just wanted to move
all the hard-coded ZIDs to symbolic constants. This is more an
experiment and a suggestion and probably shouldn't be merged.

Before I clean this one up properly, I first wanted to check if
this has consensus.

Bug: T260318
Change-Id: I680015198b289746b95872dccf0d284529875f30