ApiZObjectFetcher: Make `language` non-default, rather than allow false
The languages keys are strings and when validating them, the lone bool
(for default) will fail in Wikimedia\Message\ScalarParam which does not
accept bool.

Instead, just mark the parameter as not required and don't provide any
default, so it wouldn't even be added in the request if it were not
explicitly given.

The default is now nothing (null internally). No language will be added
in the request though, unless the user explicitly selects one.

Alternatively a default "false" string can be provided but that just
seems confusing as it could appear to be like another language.

Also added missing message for the parameter.

Bug: T265034
Change-Id: I1e61302a278de6940414c7d45693e6e83e945183