Rewrite ZObjectUtils::canonicalizeZRecord()
Instead of constantly switching back and forth between
object-world and array-world (notice the many get_object_vars()
calls, one of them in a loop), this approach does the
object → array → object conversion only once:
* Convert to array.
* Trim keys.
* Check for Z1K1 element.
* Replace local with global keys.
* Sort by key.
* Canonicalize all elements.
* Back to object.

I did a quick benchmark, and no, this is not really faster
(just a little bit). But I hope it is easier to read and to
maintain.

Change-Id: I026d9fdc60630755908b41e3e520819e6e61ab0a