Replace TypeSelector with SelectZobject
This starts the replacement of the TypeSelector component which uses
hardcoded types with the SelectZobject component with a type parameter
which fetches them dynamically from the wiki. The ListValue and
OtherKeys usages have been replaced; the FullZObject component still
uses the old TypeSelector, as the usage there is slightly different.

Bug: T264828
Change-Id: I28fd529e1631189f9ec8430e93451013c66aed62