Replace custom selector with base AutocompleteSearchInput component
SelectZobject uses now the AutocompleteSearchInput component, with the
following behavior:
* When string input is typed into the field, after `lookupDelayMs`, it
calls the `wikilambda_searchlabel` API and fills lookupResults with the
available possibilities.
* The prop `type` limits the type will be passed to the
`wikilambda_searchlabels` API to filter the types of objects to select
from.
* Emits the event `input` when a valid option is selected by:
- Searching and selecting, either by clicking on the option or
navigting with tabs and pressing enter.
- Typing the desired ZID and pressing enter or focusing out of the
input. Before selecting a ZID enter by the user, checks that it's valid
and that it exists.

Bug: T270295
Change-Id: I173d8e8ab6a1c99508a5023caed68d159421b87a