Exclude selection of Z1, Z2, Z7 and Z9 from type selector
As discussed with Denny, it would not be advisable to allow the user to
select ZObject, ZPersistentObject, ZFunctionCall or ZReference when
using the selector as a type selector (e.g. prop `type` restricts the
autocomplete to fetch only objects of the type Z4). This patch prevents
the selection of these when:
* The user types in the search field and the dropdown is populated with
available types.
* The user directly types the ZID of the type (e.g. types Z9 and presses
enter).

Bug: T270295
Change-Id: Ib5903197323732db82ba18df5b96350feca62375