Vue: Add component ZObjectGeneric and centralize typed components and mutations in ZObject
This patch involves a lot of code going from one component to another:
* Created ZObjectGeneric, which contains what before was the template of
ZObject (ID, type and a ZObjectKeyList) and is now rendering the
representation of a generic ZObject
* ZObject now handles all mutations and its template is as simple as to
decide which component to render depending on the type.
* ZObjectKey now only renders a key ID, a type selector and, when the
type is chosen, a ZObject.
* ZList renders a list of ZObject, irrespective of their types. The
selection of what component will be shown for each item is a decision
that now only lies on ZObject.

Bug: T273519
Change-Id: I77f0c71dbb709c17e1987c90a5ab71d6e2a28637