Vue: Render ZReference using ZString component
Because of the getZObjectType function now identifies as ZReference any
string that matches the regex of a Zid, we need to tell ZObject to
represent this with the same component as the ZString, or else it will
be rendered using ZObjectGeneric.
This is temporary and only until we have a ZReference component that
presents and allows edit of references in an ad-hoc way.

Bug: T266292
Change-Id: Id4710c6cbbbada42a28b4069af7643199cee616d