Vue: Apply refactor principles to ZMultilingualString
This is the last patch of the epic refactor task. Changes are:
* ZMultilingualString to emit events to let ZObject handle the state,
instead of mutating directly its input prop.
* Capture the input field `change` event instead of `input` event, to
emit the event after the value has been accepted by the field, instead
of emiting per key typed.
* Emit only the changed value, not the whole zobject to limit amount of
information passed to the parent component to only the necessary to make
the required changes.
* Change name of prop from mlsObject to zobject for homogeneity with the
rest of components.
* Removed unused getter.
* Added JSDoc.

Bug: T271878
Change-Id: I6ed0ea237db0e569ac0d2f399b79cb6af6e405df