WikiLambda.namespaces.php: Restore $namespaceNames variable
It seems to be explicitly default in mediawiki/extensions/EventLogging,
as example, so I'm assuming this is safe to do. In other words,
the LocalisationCache is not running a single variable through all
files, but rather is extracting an array from each one. As such,
this should be safe to declare explicitly.

Change-Id: If40e9803f0db74f75904bbf1cc02ce6ab7dde94a