Allow users to make API calls to the function orchestrator
This allows for evaluation of functions, though as the orchestrator
does not yet route calls down to the evaluator this only operates
for function calls using the built-in functions defined there, for
the time being.

Users can evaluate a ZFunctionCall (Z7) object either through the
API (via the new wikilambda_function_call endpoint), or inline in
the Vue interface by triggering the "Call Function" button when
creating or editing a ZFunctionCall object.

Bug: T260321
Change-Id: I0a951ff442d35762a2411d8c82a57a4217381e27