Vue: Add a Zid getter in the zobject state module
Because of the changes between canonical-normal form in the front-end,
the action was sending the whole ZReference object as the zid parameter
to the api, which was causing a failure.

Added a state getter to the zobject Vuex module that checks that the zid
being passed is the zid string value before making the API call.

Bug: T280160
Change-Id: Ie76d33ee26a84607bdb12a962916301c07171881