Normalize Z0 as a null Reference instead of as a String
Currently the front-end is receiving inconsistent values for Z2K1, while
it expects it to be a reference, when creating a new ZObject, because
it's not a valid ZReference, it's normalized as a string.

This patch adds the condition to the
ZObjectUtils::normalizeZStringAndZReference method, and it understands
as a reference both values that have valid Zid or "Z0" (used as a null
reference).

This patch also removes unnecessary override of Z2K1 value in the
front-end and fixes tests to understand normalization of Z0 as that of a
null reference.

Bug: T280513
Change-Id: I414747f550cafc342b80db721cd2ed5808aa823d