Created core error types
This implements the error types as list on
https://meta.wikimedia.org/wiki/Abstract_Wikipedia/Reserved_ZIDs
Not that this list is neither complete nor do we expect all of
them to be used. This is more like a list of error types that
can be used as a template for other error types. We will delete
error types or add new ones as needed.

Bug: T278359
Change-Id: Ibe5a5d3fa65dd2fd685691779ef617064c0dcfb1