Change language codes to ZLanguage (Z60) Zids
This patch changes all ZMonoLingualStrings Z11K1 keys from the string
representation of the language code to the Zid of a ZObject of type Z60
(ZLanguage) with all the implications of this:

* Back-end:
  * Add a ZLangRegistry to handle language code to zid translations and
viceversa.
  * Builtin ZMonolingualString and ZMultilingualString classes now
handle Zids instead of codes, and use ZLangRegistry to validate these.
  * Add maintenance and development scripts to handle the transformation
of all persisted and builtin ZMonoLingualStrings and replace their
language codes with Zids.
  * APIs that receive language codes now translate to language Zids
before querying.
  * Fix all json files.
  * Fix all tests that used language codes, manually registering with
ZLangRegistry all those language codes that were necessary for the
tests.

* Front-end:
  * Extracted all language related state, actions, getters and mutations
to languages.js Vuex module.
  * Now querying wikilambdasearch_labels to get full list of languages.

Bug: T279114
Change-Id: Iee4417863006a4a44084f29f69bd45a99f65edf1