Add optional Aliases to Z2/Persistent ZObjects as Z2K4
Aliases are conceptually maps of Z60/Languages to sets of
Z6/Strings. Since we don't have maps yet, we use the bespoke
Z32/Multilingual stringset, and since we don't have sets, we
use Z10/Lists.

This is a middleware-only change, and there is no special-
handling of these in the front-end Vue code in this patch.

Bug: T285366
Change-Id: Id38ced76bff044edb35b664eb36a3f75e253f0ac