Add SpecialObjectsByType
This creates a generic ZObject lookup by type page; functions can be browsed
by selecting the link for Function (Z8). Currently the results are presented
as a vertical list; styling into columns in some responsive way would probably
be a good idea. Paging when there are a large number of results would
also be an improvement.

Bug: T286561
Change-Id: Id267c9798d92299aa951c1e61b32506c42aa75bf