Decouple validation and ZObject creation to use function-schemata validator
This patch does *a couple* of things:
* Rewrite ZObjectFactory so that instead of creating ZObject instances
at the same time that we validate them, first uses the schemata to
perform structural validation, and if succeeded, recursively creates the
tree of built-in ZObject instances
* ZObjectFactory::create: Main endpoint that performs schemata
validation and then proceeds to construct the objects.
* ZObjectFactory::createPersistentContent: Method called from
ZObjectCOntent, calls create and makes sure that the output is a
ZPersistentObject.
* ZObjectFactory::createChild: Called recursively to construct the
ZObject instances, without being concerned about validating the input.
* The previous main functions to validate and create
(validateObjectStructure and validateKeyValue) are now deprecated and
will be removed in a future patch.
* Introduced Validator classes that create schema, validate, save the
validation output and parse it to generate ZErrors
* ZErrorFactory handles everything related to identification and
instantiation of ZErrors.
* Adaptation of all ZObject classes to be properly constructured with
the new system. And of course, their tests.

Bug: T287734
Bug: T284672
Bug: T287735
Change-Id: I47dea0901f766c00894f502f2e0b4b857823dff4