Vue: Tell MW only to run us on ES6 clients
This drops IE11 support slightly ahead of Wikimedia-wide plans, but
in practice there are going to be very few such reader/editors when
Wikifunctions launches. Doing this now allows us to write code in a
much simpler fashion, and share code between the front-end and back
-end repositories.

This disables the `no-var` eslint rule for now, as the migration to
instead use const/let will be slow and bumpy, and there's no rush.

Change-Id: Idbc3890f55c87ab3172400d8d39da9204745dde2