tests: Check that each language code known to MW has a ZNaturalLanguage
Note that this doesn't stop WikiLambda from supporting language codes that MediaWiki
doesn't yet know; we're checking that MediaWiki language codes are a sub-set of ours,
not a match. Where our language code is the 'proper' ISO/BCP47 code, we should still
match on MediaWiki legacy codes due to the alias codes in the optional Z60K2 field.

This is in 'standalone' as if this fails we'll need to change it in function-schemata
first; if we put this in the regular test it would mean that adding new languages to
MediaWiki would not be possible until WikiLambda supported them first, which is a bit
aggressive right now when we're not in production yet. In future, we may wish to re-
visit this decision.

Bug: T290593
Depends-On: I2c5789fa96321400b270e91ba07a838a8e46bc6f
Depends-On: Iff9ceff6a4764a578963d1b4195be26e46dab086
Change-Id: I6617eca16e6fa92bda2f905f9be9943b4fc4a71c