Fix canonicalization and normalization methods
This patch fixes canonicalization and normalization methods from
ZObjectUtils so that an array is understood by default as a typed list
instead of a Z10.

Added tests to ZObjectUtils (all normalization and canonicalization
related methods)

Also deprecated hybrid form (normal form with canonical lists), which
should be completely removed from code and tests in a following patch.

Bug: T298126
Change-Id: If85ca3054691b2b6b6e14c2a514cedef3cdb2600