Fix overlap between type mismatch error types
The following error types needed clarification:
* Z542: ZObject type mismatch: We are expecting a ZObject of a given
type and we receive one of a different type. (E.g. we receive a Z4 when
we are expecting a Z8)
* Z551: Schema type mismatch: Static validation error, we are expecting
a json object of a given type and receive that of another type. (E.g. we
find an array where we are expecting a string)

Fixed:
* Naming of Z551-related constants to avoid confusion
* Creation of parameters for Z551 (strings) and Z542 (references)
* ZObjectFactory should throw Z542 instead of a Z551

Bug: T303188
Change-Id: I21f817215089f87cae5fa14a301c51782d447690