Add --force and --clear options to realoadBuiltinData
This patch provides the reloadBuiltinData with a bit more flexibility
when loading the builtin data definitions in case the simple overwrite
is not sufficient (complex data migrations). The following flags have
been added to the script:
* --force Forces the reload by clearing all the secondary data and doing
an orderly insertion so that the dependencies are added first.
* --clear Also clears all the non built-in ZObjects. It must be used in
addition to the --force flag.

Bug: T311641
Change-Id: I5c94fdc8a2be869abff0350ade7953b9a185c35e