Fix cursor bug.
Prior to this fix, every new keypress in the code editor would send the
cursor to the end of the code. This was happening because:

1. The store is updated with the current code value on every keypress
2. The editor was being set with a new value every time the code value
   in the store changed.

This commit prevents this bug by ensuring the editor is only given a new
value when first initialised, or when the language changes (and so the
contents should be overwritten with the stub for that language).

This commit also corrects some existing test cases which were checking
for the presence of components by doing e.g.
```
expect( wrapper.find( 'div' ) ).toBeTruthy();
```
The wrapper is always truthy, whether the component is there or not.

Bug: T314314
Change-Id: I7f150f2794323820073920858776a4ddd97054c8