Remove default string for argument labels.
The user should be forced to specify a label themselves, making it more
likely they will choose a more informative label than 'input'.

This change causes the default value to be removed from the input box in the function-editor UI, revealing the existing placeholder text instead. In the zobject-editor UI the default value is also removed, but there is no placeholder to be revealed. Given the zobject-editor UI is considered deprecated (for functions), I am not prioritising adding a placeholder there for now.

Bug: T312120
Change-Id: I517ef7e0652f6d0f10d422d7cbfdb08d98c29936