Unify format of internal keys when they are integers
When a zobject is transformed into a table, one of the values of each
row is key, which will contain either a string with the key for the
value described (E.g. 'Z1K1') or a number with the position of the value
inside an array when the zobject is an item of a list.
This number is sometimes added as a string (E.g. '1') and sometimes as
an Integer. This is causing a number of errors all around.

This patch attempts to unify the representations of key and always
cast the value to string.

Bug: T314875
Bug: T314872
Change-Id: I0431f98fba975582585dffe231718563eed56653