Remove the setting of lookupResults, which was breaking the Type creation key selector, by 'emit'-ing a null selection.
The slight regression with this change is that options don't immediately
show on box focus, only on first type. We might want to address this
again after this fix, if it's wanted.
UPDATE: we have decided this is behavior we want, so RPCs aren't
triggered on clicking around the page.

Bug: T316547
Change-Id: I886a9ff63850799f6c5e13781bcdcaeffe1b7e2f