Update integration tests with fully-resolved built-in functions.
This is a consequence of https://gerrit.wikimedia.org/r/c/mediawiki/services/function-orchestrator/+/827541.
The changes to example-generic-(list|pair) are particularly unfortunate, as is
the deactivation of the test involving generic maps. It
turns out that our requests are now too large for the function-evaluator.
This has been captured in a bug
(https://phabricator.wikimedia.org/T318363).

Bug: T303580
Change-Id: Ie200a4926fed401964bfe9516c7e4fc979ab359a