Fix label-dependent data model behind the function page. Part A.
1) Make currentZObject.js' "currentZObjectLanguages" the source of truth
for all languages associated with a function. This includes languages
found in the name labels, argument labels and alias labels. This ensures
if the function editor is opened, each language is editable
if it has at least one of those label elements.

2) Set the scaffolding for alias labels to be an empty benjamin array
(with just the type), rather than an empty alias stringset set to the
userLang.
- This initial default empty stringset set to the userLang had a bug and
wasn't getting updated on language selection changes. Spoke
with AdesojiThisDot@ and we agreed that the scaffolding was perhaps not
necessary for alias' and could be removed. If someone else disagrees,
please feel free to comment!

Bug: T318143
Change-Id: Ic1b09ba15f44d4855d9477fbc072dbe7351aa49d