Fix editing a ZType: not possible to change the value type of an existing key.
The issue stems from the fact we were passing the zType of the key value
as a fallback, if the literalType of the key itself was not available.
For example, the literalType of a Z3K1(Key: value type), is a Z4 (Type).
However once a value is selected, that value could be a Z9 (Reference),
like "String". This Z9, was being passed as the type to the
ZObjectSelector and used for the getLookupResults type, however, we
cannot lookup on a Z9 reference.

The solution I have here is to therefore remove that fallback. If it is
an existing zobject we haven't run fetchZKeys by the time the selector
loads. So now the ZObject key will first check if we have the type
literal, and if not, fetch it, before passing it to the
zObjectSelector.

Bug: T320277
Change-Id: I721b4a468988b84b343217b254f01fde04a50940