And the "confirm" dialog to the function editor when a user attempts to click away from the page after making edits and before saving.
The shortfalls of this patch:

* It covers clicks of links on the page that navigate away, but does
not include the back button/page refresh. Because we have built our
own router (instead of Vue Router), I didn't have access to certain
built in navigation guards, and I couldn't figure out hooking into our
own router.js to get it to work.
Possible option, use the built-in browser 'are you sure' dialog for
those two actions if we really want to give guard losing changes that
you get from window.onbeforeunload?

* "Dirty" for the labels is if you ever changed the inputs. It doesn't
become "clean" again if you go back to the label's original state. I did
implement this for the names and arguments labels, but it was quite a
lot of code, and I hadn't even got to aliases yet. IMO I think it's not
enough of a gain compared to this option.

* There is no integration test covering the dialog as we only load the
Vue app, and all links are outside of that.

Steps to test:
1. Go to the function editor, and view the publish button is disabled.
2. Attempt to click away from the editor via any link on the page, and
find you are able to.
3. Go back to the editor and make a change to any label, input or output
type.
4. See the publish button is now enabled.
5. Attempt to click away on any link and get the 'are you sure?' dialog.

Bug: T321381
Change-Id: I0c48b6bb1a3fdb83cdc4ee92c603b3365970cdf6