Setup zObject editor "isDirty", which is connected to the Publish Button being disabled or active.
Each action will need to be included as part of the ZObjectEditor
integration tests once we have them (if we want that coverage).
I have added unit tests for now.

To Test:
-Go to the zobject editor (either existing or new zobject) and see the
publish button is disabled.
-Make any change in the editor (try all you can come across if
possible!) and see that the button becomes active.

Next up: Connect to the "are you sure" dialog on page exit.

Bug: T321381
Change-Id: Id3e9f00229e375758138ac46d772692d3165c65b