Fix injection of expected and actual values into perform_test result.
This was broken. The PHP code only injects these values if the
validation result is false. Previously when checking whether this was
false it was expecting the validation result to be a Z40 with its Z40K1
set to "Z41" or "Z42" directly; actually the Z40K1 is set to a Z9
pointing to one of the two.

Also add assertions that this data is injected correctly to
ApiPerformTestTest.

Bug: T312290
Change-Id: I7b2a790863cc1b19b19faa354b25d492252d88e5