Fix inappropriate use of normal-form Strings on Z3K2s.
The regex change in schemata.js ensures that key Ids (containing a 'K')
do not set the normal-form string object, but rather just the String
value.
The slightly strange behavior with this now though is if the user
removes the "K" from the keyId whilst editing, the z-object jumps to
normal form (as it updates in real-time).
However, as the correct form for a keyID is <RootZobjectId-K-KeyIndex>,
perhaps the question is whether we should allow editing of these key
IDs at all?

The other change moves the setting of the keyID (eg Z0K1) from the ZKey
component to the initial zObject creation. Previously setting it in
the ZKey component had two issues:
1) Because of the '!value' check, it meant it pushed a new value to the
zobject tree if the user ever cleared the box, which meant the next
keyId would suddenly populate (eg Z0K2).
2) The zobject in the json tree became out of sync with the state of the
component (and therefore didn't update), because the ID had increased
and therefore didn't align.

To test:
Go to wiki/Special:CreateZObject
Create a Type
Add a Key
In the "currentZObject" JSON tree see that Z3K2: Z0K1
Update change the Z0K1 in the editor and see it update.
Add another key, see it is pre-populated with Z0K2.

Bug: T316793
Change-Id: I352d6aee5536a94cbc5ce2172982a6010658b99f