Avoid race condition by having initializeZObject return promise.
In App.vue we are careful to await initizalizeZObject before evaluating
the URI, to ensure that the ZObject is initialized before we start
building the UI to edit it. But in the case of a new object,
initializeZObject was not actually returning a promise to be awaited -
this led to it building scaffolding for a new object *after* the
function editor has already been loaded, which led to the function
editor manipulating different local state to that used when publishing.

This bug wasn't spotted by our integration test unfortunately - I
suspect this is because the test environment executes async work using a
mechanism that is different in some way to production.

Bug: T325685
Change-Id: I9778e6dce1131202bf9b8d8af13cd5b9befcc3b0