Fix deleting then re-adding alias labels.
With this change, when the last alias label is deleted, the empty alias
scaffolding block (the Z31 in the Z32K1 list) is deleted. This resets
the zobject tree to its original state. Previously, only the alias
itself in the Z31K2 list was deleted, which meant when a first alias was
re-added, a whole new Z31 block was incorrectly added.

ZLabelBlockAliases will not exist as it does now once the default view
is implemented. There are currently no unit tests for this component so
I'm not sure it's worth adding a whole set for this change?

Bug: T325973
Change-Id: Id0ec2d3098c60ee0906186552162fe125a004701