OrchestratorRequest: Memoise Wikifunction functions calls in memcached
This changes the return from the orchestrate() method from a PSR-7 wrapper
object (which is short-lived) into the final string representation (which
is what we were using in each case anyway), as the cache is shared across
the cluster and over (potentially extensive) time.

Facility remains for the non-cached method, for use by ApiHealthCheck, as
otherwise the health check would only be testing the MediaWiki/memcached
part of the system, and not the actual orchestrator/evaluator part about
which we actually care most.

Follow-up TODOs:
* (T338244) Tests of using the non-MainStash cache
* (T338245) Set up the key better for changes to implementations
* (T338246) Explore simplification of the key

Bug: T307742
Change-Id: Iebe021f4f4a38d38198d61db2483c39d4cad551b