Allow greater flexibility re: when implementation re-ranking occurs
Background: implementation ranking involves a function's attached
testers and implementations, and considers all of them.  To do the
ranking, we need test run results covering all of the attached testers
and implementations.  The existing behavior is to do the ranking when
no testers or implementations are specified in the input parameters
(because in that case we automatically retrieve results for all attached
testers/implementations).

This patch allows ranking to occur when testers and/or implementations
*are* specified, so long as those specified include all of the attached
testers/implementations.  (Here again, we will have all the results needed
to do the ranking.)

If any non-attached testers and/or implementations are included in what's
specified, we ensure that only the *attached* testers & implementations are
considered in the ranking.

Also, we remove ApiPerformTest parameter updateImplementationRanking.
(Henceforth we will always do the update, if certain conditions
are met.)

Bug: T319478
Change-Id: I79fdf927ceacedef7e4f298c9ec1a31837ddaa93