Update use of CdxButton `type` and `action` props
In Codex v0.7.0, the CdxButton's `type` prop changed. The `type`
prop was renamed `weight` and will continue to accept the same
values ('normal', 'quiet', and 'primary'). A new `type` prop was
added that accepts the old types, plus native button types
('button', 'reset', and 'submit'). In a future release of Codex,
the `type` prop will be removed altogether and the native type
can be bound to the CdxButton via the `type` attribute.

So, while the old values for `type` still work, we want to
update the prop name in anticipation of the `type` prop's
removal. For any new code moving forward, the `weight` prop
should be used.

This patch:
- Renames all uses of the `type` prop to `weight`
- Changes a few instances of `:destructive="true"` to
  `action="destructive"`

Bug: T312987
Change-Id: Ifcd6380e4af9104dfec375a77927a4f421c3a22e