PageRenderingHandler: Let /wiki/fr/Z901 transparently give the user the content in fr
This uses MediaWiki's entry-point routing, but some in-context links are not yet
re-written by this change to use this pattern, notably our own inside WikiLambda
but also e.g. on wikitext pages, so the first click after arriving will still be
one that takes users to an English-language page as-is.

Another thing to do is dump into the <head> a bunch of links of the form:
  <link rel="alternate" hreflang="$1" href="$2">

Bug: T268678
Change-Id: I819822eb0430b71bc7ff469c3be122110f9a512c