Metrics custom data: pass undefined as null & update 2 names
Wherever there's an undefined value for a Metrics Platform
custom data property, pass it in as null.

Explanation:
Metrics Platform accepts a null value for a custom data property, but
does not accept undefined.  If it gets undefined, no event gets
propagated.  Anytime there's an undefined value for one of these
properties, it will be more useful to have the event produced.  For
example, sometimes zobjecttype, in DefaultView.vue, will be
(properly) undefined and we need an event showing that.  To capture all
such cases,  we make them pass null everywhere.

Also, for consistency of custom data property names,
  change contenttype to zobjecttype, in one place
  change contenttype to implementationtype, in one place

Bug: T340784
Bug: T340160
Bug: T340779
Bug: T340783
Bug: T340786
Change-Id: I6ed9ce44893bd99daffc14fb1dda76425df11246