Special:CreateObject: Restrict rights to 'wikilambda-create' not 'createpage' right
… and actually check and display the error message here, and for
Special:RunFunction too. Oops.

This alters the browser test expectations, to give an error unless logged-in.
Note that the changed behaviour for RunFunction does not affect the browser
tests as they run in our default configuration, in which even logged-out
users have wikilambda-execute rights, so the 'Permission error' screen will
not show up here, but will in production as currently configured.

Bug: T342859
Change-Id: I5a25b64503a151b522a4da7c9768a158601a6b22