PageRenderingHandler: Set href always, even when we don't set the label
When this method was originally written, it was purely for re-writing the
label, so it was fine to stop processing early; now we re-write the href
always (as long as it's "ours"), in any context, so re-order the code and
add comments to observe what happens when.

Add tests so this doesn't recur.

Bug: T343041
Change-Id: Iec2cb10a6ed61d14e7abff8e9cda6e203211bb6c