Add permission checks for bool, unit
In ZObjectAuthorizationInCreationTest add a work-around because these
kinds of creations are now doubly-blocked, in the user-rights system,
and also in the structural validation system, so instead of using the
whole test rig, bail early if "the system should stop this". In cases
affected this way, if the structural validation changes, we can relax
this and run the normal test suite instead.

Bug: T349497
Change-Id: I2453fea913220d32e42fe3b26f6178ca553bbb77