Rename zKeys module to library
zKeys module and related state objects and methods are badly named.

The module was initially named zKeys because it only handled the keys of
zTypes, but it has evolved ever since.

This module currently fetches and stores all the zobjects that are
needed by the UI, and the labels of the objects, of type keys and of
function arguments.

It also contains helper getters to consult properties of stored (not
root) auxiliary objects (labels, key or argument names, language codes,
function details, etc.)

Bug: T329107
Change-Id: Ib5a229f52335b4ff534dce45fdec8dc52cd888cf