Set isDirty only when making changes in main object, not on detached objects
The Function Editor widget sets in the zobject table two different
detached objects: one for the function call (editable) and another for
the function call result (non editable).
These detached objects are also represented with ZObjectKeyValue.

When making changes in the editable detached object (function call input
fields), we should not set the page as dirty (with changes)

Bug: T344169
Change-Id: If15860c737ea2620e7e69c75c7f92d659133ed03