ZObjectEditAction: Use getTitle() not getPageTitle() to get the target page
This was always conceptually broken - getPageTitle() used to return either a
string or a Message that was used to put into the <h1> of the page, but since
5c8be3162876a78a9147a4869e5c56aacbd1fd65 landed in MediaWiki this is always
now a Message object, which was breaking our code.

Instead, use the actual Title object for the page via getTitle(), as we do in
ZObjectContentHandler and SpecialViewObject already.

Bug: T347528
Change-Id: I24fd2deb51c4bff1229844bf3f03b1fa28b06454