Remove multiple instances of MetadataDialog component in FunctionReport widget
Context:
A FunctionMetadataDialog was created for each entry in the table, but
the dialog is being treated like a single-instance component.
This means that whenever we click "Details", we configure the dialog
with whatever info it must contain, and we open it.
However, because the dialog was inside the v-for loop, there were n
dialogs being opened and overlayed instead of just one.

This fix simply takes the component out of the loop (as it was
intended).

Bug: T346972
Change-Id: I0b37884105c65ccad7686b6a24aa8631470130ee