styles: Use new forward-facing breakpoint variables
Replacing legacy breakpoint variables with new Codex
design system tokens, here mostly `@max-width-breakpoint-mobile`.
This also makes the code a bit easier to read.
Note, that the values have slightly changed in accordance with design
and Web team. We are bringing the new breakpoint values to all skins
and extensions.

Also using 'mediawiki.skin.variables.less' as reference as it enables
skin specific overrides of the Codex tokens in default 'WikimediaUI'
style as provided to Vector and MinervaNeue skin.

Bug: T331403
Change-Id: I950024c6acd5a088fbb2dca74ef54c49d190a36e