build: Upgrade eslint-config-wikimedia from 0.25.0 to 0.26.0 and make pass
Whilst we're at it, upgrade grunt-eslint from 24.0.1 to 24.3.0, and drop
our secondary load of eslint-plugin-jest (already provided by our config).

To pass against the updated Wikimedia-wide coding standards, this makes a
large number of auto-fixes, mostly for template-curly-spacing. We're also

Also, the new whitespace means the lines-covered rule has gone down, so to
pass this adjusts the threshold down.

Merge the vue/no-undef-components and vue/no-unregistered-components rules
as the former has replaced the latter; drop all the old components we have
since dropped.

Change-Id: I119a8a49d87aa3885f64215b1a547acc23862008