Fix fallback into link when the type has been seen
Current implementation only falls back into reference when the type has
been seen a second time.
This means that with identity fields (e.g. Z46K1 or Z64K1) we are
creating a literal as the identity and a reference for the second level
identity field.
Instead, we should be creating a reference for the first level
recursion.
This patch also simplifies the keyList from a chained object into an
array, as it only registers a list of types that have been seen before
and does not care for the position of those.

Change-Id: Ib63e8a8264e07124e7514602fded26456f823294