Metadata dialog: handle new debugging metadata from executors
This includes the key executorDebugLogs.  It can also include
arbitrary keys inserted by user code.  In all cases, the
values are lists of strings.

This patch provide internationalization for executorDebugLogs, and
formats each list of strings as a bulleted list.

Reducing 3 coverage thresholds for mixins, because portray.js
is likely to be reworked per T341127.

Bug: T353544
Change-Id: I4cf74b81a5270f3c5c99bae4def2315508ede822