ApiZObjectFetcherTest: Don't write to historic property to avoid PHP 8.2 being sad
titlesTouched used to be a thing, but was dropped a few years ago,
after which this was a silently-pointless creation of a dynamic
property. However, with PHP 8.2 and above, this is no longer silent
but instead triggers a warning, so let's stop doing it at all.

See I09fd4618e955d814e24f3b8781bc92487f11287a for the main change.

Change-Id: Ibfd5a422ff18c21413d0489b7f64fd9a8ffffce2