Move browser-test running commands to 'npm run browser-test' etc.
This allows us to point the 'npm run selenium-test' reserved entrypoint
at a different thing than what we want to actually use as humans, or in
DUCT/etc. proper test runners, so we can resurrect the gerrit CI normal
'selenium' job.

Whilst we're at it, adjust some of our language terms for consistency.

Change-Id: I4730da1141f0035abf8ef8e2160246db2b09895e