Cleanup: Use zobjectToRows instead of the raw Row objects whenever possible
To feed the store zobject table we can directly assign the array of rows
or use zobjectToRows helper method, which uses tested transformation
mixins.
Sometimes it is more clear to use the row data directly, but often we
want to make the tests a bit simpler and just be able to set the store
with the ZObject value.
This will hopefully make tests much more maintainable and trustworthy.

Change-Id: I70888e0ce2485b27d7b9dabc81a35b12142a3e4d