Cleanup: Change non-terminal row value from string to integer constants
Row value field contains either a terminal string, or a value that
indicates that the row is non-terminal and parent to an object or an
array.
Before, the object or array rows were indicated using string constants,
but this is problematic because we could not distinguish an object row
from a terminal row with the value 'object', and similarly with arrays.
For this reason, the non-terminal rows need to be marked with non-string
values.

This patch does:
* Replacement of all literal 'object' and 'array' with their constant or
the methods row.isArray(), row.isObject() or row.isTerminal()
* Set the constants to something other than a string.
* Make sure that tests work!

Bug: T329107
Change-Id: I780dc280c6d0a9e684756da138be9cdf8a90afeb