Bugfix: Pass keyList by value when creating new argument paths
createObjectByType controls the repetition of types that have references
to themselves by observing in depth the times a type is created, and
only allowing one level, while creating an instance of reference on the
second reference.
For this, when creating new createObjectByType branches, keyList must be
passed by value, instead of by reference, so that what occurs in one
branch does not affect the others.
This must be done in depth, but when this repetition occurs from
parallely occurring arguments, it must not be restricted.
Added regression test.

Change-Id: I193760cde75854bbab5523141d591b6d3a47e6d8