store: When transmitting inline code, encode pipes so they make it
MediaWiki uses the pipe character in its API calls to distinguish
between different values. Unfortunately a pipe character is pretty
important in code, no least in the context of binary operations, so we
need a way around this. As a short-term measure (at least), let's
encode the '|'s as '🪈's, which are much useful and so (we hope) a lot
less likely to be used in meaningful code (as opposed to values, where
they are still fully working, or in comments, where their alteration
doesn't matter).

This is imperfect, but better than the current situation!

Bug: T358089
Change-Id: I69da217eec08e94a61102e25ec8c37b8997d59bd