ServiceWiring: Use `newFromParams()` via MW services
The code is already in the right place meaning we can just access
the service container and get an instance of the cache.

`ObjectCache::getInstance()` bypasses service wiring and we want to
fix cases whereby this is not needed. The new ObjectCacheFactory
wants to unify access patterns as much as possible so we're sure all
our codebases are consistent in how they do things.

Bug: T358346
Depends-On: I3179a387486377c6a575d173f39f82870c49c321
Change-Id: Ic3cf6a5575c887163bdf50045799ddd7dd7bc1d3