Ability to directly 'publish' an old edit
When we visit a previous edit and wish to revert to that, it seems
our Publish button is defaulted to 'disabled' requiring us to make
some change for us to enable it. We want to avoid having to do this
by already enabling that button when we visit such a page, which
would have a query with a valid 'oldid'.

Bug: T343654
Change-Id: If7a96e3c836cdaf767fd47100ea62937175c5b10