ApiFunctionCallTest: Disable on old release branch, services have changed
We don't support these old branches, but at least let's get the
tests back to working by disabling the doomed-to-fail tests, as
we don't run old versions of the services in Beta Cluster, only
the latest.

Also adjust the expected error returned by  one of the cases in
ApiPerformTestTest, as it's simple enough a "fix".

Change-Id: If92c80eb1df80ce0dab4f009727bcd24b0aa4118